requirements.txt
beautifulsoup4
requests
openai==0.28.1
python-dotenv
streamlit
fpdf